refactor: share operator user login method helper
